# Use BMI2 when the CPU supports it for certain field element operations