--- a/LICENSE
+++ b/LICENSE
@@ -1,6 +1,7 @@
 Copyright (c) 2009 The Go Authors. All rights reserved.
 Copyright (c) 2017 George Tankersley. All rights reserved.
 Copyright (c) 2019 Henry de Valence. All rights reserved.
+Copyright (c) 2021 Oasis Labs Inc.  All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are