--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,5 @@
 module github.com/gtank/ristretto255
 
 go 1.11
+
+require golang.org/x/sys v0.0.0-20210124154548-22da62e12c0c