Square is almost not worth it, though it may be worth investigating if it is possible to use ADCX/ADOX to make both Mul and Square faster (Isn't registry pressure fun?).  Unfortunately the stub ends up being somewhat ugly, but having to fight with the inliner shouldn't come as a surprise at this point.

```
name \ time/op  /tmp/baseline  /tmp/mul-mulx  /tmp/square-mulx
WideMultCall-8    3.04ns ± 0%
Add-8             14.5ns ± 0%
Mul-8             36.8ns ± 0%    31.6ns ± 0%
Mul32-8           14.0ns ± 0%
Square-8          29.0ns ± 2%                      26.5ns ± 0%
```