@FiloSottile 2021-02-09
Hi @Yawning! Two quick questions.

First, we did a bunch of work on the field implementation getting it into filippo.io/edwards25519, and eventually I am hoping to replace the guts of this package with that. Would you consider contributing these changes there? I am trying to keep that repository CLA-clean to upstream it into Go, as well, so it would involve signing the Go CLA and agreeing to submit this code as a contribution to Go.

Second, do you have any higher level benchmarks, ideally through benchstat, so we can evaluate if the extra code complexity is worth it?